Begins writing tests with enzyme/jest
