+++ client/src/index.jsx
@@ -1,42 +1,6 @@
-/* eslint-disable comma-dangle */
-/* eslint-disable no-console */
+/* eslint-disable import/extensions */
 import React from 'react';
 import ReactDOM from 'react-dom';
-import $ from 'jquery';
-
-
-class App extends React.Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-
-    };
-  }
-
-  componentDidMount() {
-    this.getFirstReservations();
-  }
-
-  //
-  // eslint-disable-next-line class-methods-use-this
-  getFirstReservations() {
-    $.ajax({
-      method: 'GET',
-      url: './reservations',
-      success: (data) => {
-        console.log(data);
-      },
-      error: () => {}
-    });
-  }
-
-  render() {
-    return (
-      <div>
-        Hello From index.jsx
-      </div>
-    );
-  }
-}
+import App from './App.jsx';
 
 ReactDOM.render(<App />, document.getElementById('app'));
